A bit of a cleanup for SubSigns Create an SubSigns object instead of using a static class
